+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.8.3](https://github.com/InseeFr/Pogues/releases/tag/3.8.3) - 2026-09-17
+
 ### Added
 
 - Add a warning modal when editing a code list used by questions, warning the user that the collected variables will be regenerated.